// lib/parts/hair-styles/types.ts
import type { PartMask } from '../types'

export interface HairStyleBundle {
  head_front_overlay: PartMask
  head_top: PartMask
  head_right: PartMask
  head_left: PartMask
  head_back: PartMask
  head_bottom: PartMask
  hat_front: PartMask
  hat_top: PartMask
  hat_right: PartMask
  hat_left: PartMask
  hat_back: PartMask
}